Add statistics

+++ AdventureDayBackend/portal-api/src/team-management-api/Data/TeamScore.cs
@@ -12,7 +12,7 @@ namespace team_management_api.Data
         public Team Team { get; set; }
 
         [NotMapped]
-        public int Score => Wins + Profit - Errors;
+        public decimal Score => Wins + Profit - Errors;
 
         public int Wins { get; set; }
 
@@ -21,10 +21,10 @@ namespace team_management_api.Data
         public int Errors { get; set; }
 
         [NotMapped]
-        public int Profit => Income - Costs;
+        public decimal Profit => Income - Costs;
 
-        public int Income{ get; set; }
+        public decimal Income{ get; set; }
 
-        public int Costs { get; set; }
+        public decimal Costs { get; set; }
     }
 }
